Wrap up lexer part in thesis and remove static cast

+++ nulascript/lexer/lexer.cc
@@ -45,8 +45,7 @@ void Lexer::skipOverWhitespace() {
 }
 
 char Lexer::peekNextChar() {
-    // ? is casting needed here
-    if (readPos >= static_cast<int>(input.size())) {
+    if (readPos >= input.size()) {
         return 0; // EOF
     }
